+++ src/components/ui/PopOver/index.tsx
@@ -6,7 +6,7 @@ import React, {
   createRef,
   useState,
   HTMLAttributes,
-  useEffect
+  useEffect,
 } from 'react';
 import { Manager, Reference, Popper } from 'react-popper';
 
@@ -48,7 +48,7 @@ export const PopOver: FC<PopOverProps> = ({
   children,
   isSubMenu = false,
   placement = 'bottom-start',
-  a11yLabel
+  a11yLabel,
 }) => {
   const menuRef = createRef<HTMLSpanElement>();
   const [isOpen, setIsOpen] = useState(false);
@@ -88,6 +88,10 @@ export const PopOver: FC<PopOverProps> = ({
   };
 
   const closePopover = (e: any) => {
+    // workaround for input file component - stop from removing it from the DOM
+    if (e.target.tagName === 'LABEL' || e.target.tagName === 'INPUT') {
+      return;
+    }
     const isSubMenuButton = e.target.closest("[data-menu='submenu']");
     return !isSubMenuButton ? setIsOpen(false) : false;
   };