+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add `useLocalAudioInputActivityPreview` hook for direct access to microphone input value
+- Add Share Video File feature
 
 ### Changed